+++ Runtime/SceneManagement/GameSceneSO.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+
+namespace CustomTools.SceneManagement
+{ 
+    [CreateAssetMenu(menuName = "Custom Tools/Scene Management/Game Scene SO")]
+    public class GameSceneSO : ScriptableObject
+    {
+        [Header("Information")]
+        public string sceneName;
+    }
+}