+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this package will be documented in this file.
 
+## [1.0.19]
+Add scene management scripts and metadata
+
+This commit introduces a set of scripts and metadata files for scene management in Unity. It includes 'GameSceneSO', a scriptable object class for game scenes, and 'GameSceneSOEditor', a custom editor class for modifying scene attributes. Metafiles associated with these classes and their directories were also added
+
 ## [1.0.18]
 This commit introduces a new SingletonMonoBehaviour class in the Unity custom tools runtime. This class creates a singleton of MonoBehaviour that survives scene changes in Unity Engine. The SingletonMonoBehaviour ensures that only one instance of a MonoBehaviour exists, and this instance is reusable across the entire project.